server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["B4TRUH4 <matveybond2002@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.6"
djangorestframework = "^3.15.1"
dj-database-url = "^2.2.0"
psycopg2-binary = "^2.9.9"
djangorestframework-simplejwt = "^5.3.1"
djoser = "^2.2.2"
gunicorn = "^22.0.0"
drf-spectacular = "^0.27.2"
drf-spectacular-sidecar = "^2024.6.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
